refactor: consolidate permision constants in one place
