+++ src/authz-module/roles-permissions/index.ts
@@ -1,0 +1,15 @@
+export {
+  CONTENT_LIBRARY_PERMISSIONS,
+  libraryResourceTypes,
+  libraryPermissions,
+  libraryRolesMetadata,
+  rolesLibraryObject,
+} from './library/constants';
+
+export {
+  CONTENT_COURSE_PERMISSIONS,
+  courseResourceTypes,
+  coursePermissions,
+  rolesObject,
+  courseRolesMetadata,
+} from './course/constants';
